More support for PID tuning

+++ firmware/motor_controller/src/JsonCtl.h
@@ -69,7 +69,7 @@ class JsonCtl {
 
   void update_ctl_struct();
   void clear_name_val_info();
-  int16_t val_string_to_num();
+  int32_t val_string_to_num();
 
   Stream & ser;
   enum ParserState {NOT_IN_CMD, CMD_NAME, CMD_VAL, CMPL_VAL, CMPL_CMD, ERROR};
